+++ Semana_4/node/src/modules/users/infra/http/routes/users.routes.ts
@@ -1,30 +1,20 @@
 import { Router } from "express";
 import Users from "../../Typeorm/entities/Users";
 
-
 import ensureAuthenticated from "../middlewares/ensureAuthenticated";
 import multer from "multer";
 import uploadConfig from "../../../../../config/upload";
-import UserController from '@users/infra/http/controllers/UserController'
-import UserAvatarController from '@users/infra/http/controllers/UserAvatarController'
+import UserController from "@users/infra/http/controllers/UserController";
+import UserAvatarController from "@users/infra/http/controllers/UserAvatarController";
 
 const usersRouter = Router();
 const upload = multer(uploadConfig); // Instancia do multer
-let userController = new UserController()
-let userAvatarController = new UserAvatarController()
+let userController = new UserController();
+let userAvatarController = new UserAvatarController();
 
-usersRouter.post("/", userController.create );
+usersRouter.post("/", userController.create);
 
-// usersRouter.get("/", async (request, response) => {
-  
-//   let users = await userRepository.find();
-
-//   users.forEach((user) => {
-//     delete user.password;
-//   });
-
-//   return response.json(users);
-// });
+usersRouter.get("/", userController.index);
 
 usersRouter.patch(
   "/avatar",