+++ Semana_4/node/src/shared/container/providers/index.ts
@@ -1,6 +1,19 @@
-import {container} from 'tsyringe'
+import { container } from "tsyringe";
 
-import IStorageProvider from '@shared/container/providers/StorageProviders/models/IStorageProvider'
-import DiskStorageProvider from '@shared/container/providers/StorageProviders/implementations/DiskStorageProvider'
+import IStorageProvider from "@shared/container/providers/StorageProviders/models/IStorageProvider";
+import DiskStorageProvider from "@shared/container/providers/StorageProviders/implementations/DiskStorageProvider";
 
-container.registerSingleton<IStorageProvider>('StorageProvider',DiskStorageProvider)
+import IMailProvider from "@shared/container/providers/SendEmailProvider/models/IMailProvider";
+import EtherealMailProvider from "@shared/container/providers/SendEmailProvider/implementations/EtherealMailProvider";
+//import IStorageProvider from '@shared/container/providers/SendEmailProvider/implementations/'
+
+container.registerInstance<IMailProvider>(
+  "MailProvider",
+  new EtherealMailProvider()
+);
+
+container.registerSingleton<IStorageProvider>(
+  "StorageProvider",
+  DiskStorageProvider
+);
+//container.registerSingleton<IMailProvider>('MailProvider',MailProvider)