+++ Semana_4/node/src/shared/infra/http/middlewares/logRoutes.ts
@@ -1,0 +1,11 @@
+import { Request, Response, NextFunction } from "express";
+
+export default function LogRoutes(
+  request: Request,
+  response: Response,
+  next: NextFunction
+) {
+  let log = `[${request.method}] - ${request.url}`;
+  console.log(log);
+  next();
+}