+++ Semana_4/node/src/shared/infra/http/server.ts
@@ -3,7 +3,6 @@ import express, { Request, Response, NextFunction } from "express";
 import "express-async-errors"; // Para poder tratar errors async, deve estar abaixo da importacao do express
 import routes from "./routes";
 import Cors from "cors"; // Serve apenas para apps com interface web
-import "reflect-metadata";
 import "@shared/infra/Typeorm";
 import uploadConfig from "@config/upload";
 import AppError from "@shared/errors/AppError";